fix use client (修复在bio项目中引用时 提示需要use client的问题)
